+++ changelog.md
@@ -35,6 +35,7 @@ Codion Change Log
 - EntityTablePanel.setEditComponentFactory() and setTableCellEditorFactory() moved to Config.
 - EntityTablePanel.referentialIntegrityErrorHandling() moved to Config.
 - EntityTablePanel.refreshButtonVisible() moved to Config.
+- EntityTablePanel.statusMessage() moved to Config.
 
 ## 0.17.31
 ### is.codion.common